Updated user service's userId getter method;
EditActivate guard was not working properly, when there was not logged in user.
Not logged in users were still able to access item's edit page. Apparently the if statement was failing on the id's match. Therefore, userId getter method now returns user's ID or empty string, if the user ID is undefined;